largest-contentful-paint/progressively-loaded-image.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: Progressively Loaded Image</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
  promise_test(async function (t) {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
    const beforeLoad = performance.now();
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const imgElement = document.getElementById('image_id');
        const imageSrc =  get_host_info().HTTP_ORIGIN + imgElement.getAttribute('src');
        const imageWidth = imgElement.naturalWidth;
        const imageHeight = imgElement.naturalHeight;
        const size = imageWidth * imageHeight;
        const lcp = entry.loadTime;
        const resourceLoadTime = calculateResourceLoadTime(imageSrc);
        assert_greater_than(lcp, resourceLoadTime,'LCP event fired');
        checkImage(entry, imageSrc, 'image_id', size, beforeLoad);
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
    function calculateResourceLoadTime(imageSrc) {
    const entries = performance.getEntriesByName(imageSrc);
    const resourceEntry = entries[0];
    assert_equals(entries.length, 1, 'No Resource timing entry found');
    return resourceEntry.responseEnd - resourceEntry.startTime;
  }
}, 'Same-origin image is observable.');
</script>
<img src='/images/arrow-oriented-upright.jpg?pipe=trickle(17280:d1)' id='image_id'/>
</body>
</html>
